googlenews-fetch-news/pyproject.toml
[project]
name = "googlenews-fetch-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gnews>=0.4.3",
    "pyyaml>=6.0.3",
]
